% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSD.R
\name{VSD}
\alias{VSD}
\title{Viral Shedding Detector.}
\usage{
VSD(expr, reference)
}
\arguments{
\item{expr}{A matrix or data frame consisting of viral RNA expression values.}

\item{reference}{Simulated viral RNA detection patterns of viruses inside or outside the cell.}
}
\value{
A dataframe which contains three parts: viral replicons, viral particles and viral fragments abundance.
}
\description{
VSD is an optimization model developed using a linear programming algorithm which includes an objective function, constraints, and decision variables. VSD, considers the abundance of viral replicons and viral particles as variables in its optimization process.
}
\details{
Hi! Welcome to VSD, a tool for calculating viral replicons and viral particles abundance
based on viral RNA expression profiles.

 Input data:  Normalized viral RNA sequencing data, and the normalization should be corrected for RNA length;
Output data: Abundance of viral replicons, viral particles and viral fragments (fitted residual representation)
}
\examples{
VSD(expr=example,reference=ref)
}
